// src/components/sections/glms/types.ts
export interface GradientTextStyle {
  fontFamily: string;
  fontSize: string;
  fontWeight: number;
  letterSpacing: string;
  lineHeight: string;
  background: string;
  WebkitBackgroundClip: string;
  WebkitTextFillColor: string;
  backgroundClip: string;
}

export interface TextStyle {
  fontFamily: string;
  fontSize: string;
  fontWeight: number;
  letterSpacing: string;
  lineHeight: string;
}